refactor: extract Starfield class to its own module
- Moved Starfield class from sprites/sprites.py to sprites/starfield.py
- Added import for Starfield in sprites/__init__.py
- Verified the background starfield renders correctly
