+++ sprites/__init__.py
@@ -1,1 +1,2 @@
 from .sprites import *
+from .starfield import Starfield
